- Automatically detect destination within source directory tree and exclude it. - Prevent accidental expansion into rm -rf /... - Use globbing instead of "ls" to get backup directories.

+++ home-snapshot.sh
@@ -8,11 +8,13 @@ if [ -z "$source_dir" ]; then
 	notify-send -a "home-snapshot" -u critical 'Source directory is not specified' "Fix your configuration at $HOME/.config/home-snapshot.conf"
 	exit 1
 fi
+
 destination_dir=`echo "$config" | grep 'DESTINATION_DIR=' | cut -d = -f 2`
 if [ -z "$destination_dir" ]; then
 	notify-send -a "home-snapshot" -u critical 'Destination directory is not specified' "Fix your configuration at $HOME/.config/home-snapshot.conf"
 	exit 1
 fi
+
 max_snapshots=`echo "$config" | grep 'MAX_SNAPSHOTS=' | cut -d = -f 2`
 if [ -z "$max_snapshots" ]; then
 	notify-send -a "home-snapshot" -u critical 'Max snapshots is not specified' "Fix your configuration at $HOME/.config/home-snapshot.conf"
@@ -22,7 +24,7 @@ fi
 today_date="$(date '+%d-%m-%Y')"
 # Full backup path
 backup_path="$destination_dir/$today_date"
-#latest link is used for hard linking snapshots to reduce used space
+# latest link is used for hard linking snapshots to reduce used space
 latest_link="$destination_dir/latest"
 # Use to add the --no-perms as workaround for filesystems not supporting Linux permissions
 no_perms=`echo "$config" | grep 'NO_PERMS=' | cut -d = -f 2`
@@ -33,19 +35,29 @@ else
     no_perms=""
 fi
 
+# Detect destination within the source tree to exclude it from rsync
+exclude_relative=`realpath --relative-to="$source_dir" "$destination_dir"`
+
+if [ -z "$(echo "$exclude_relative" | grep "^\.\./")" ]; then
+    exclude_relative=`echo "--exclude='$exclude_relative'"`
+else
+    unset exclude_relative
+fi
+
 log_location='/tmp/home_snapshot.log'
 
 #If a backup was taken today, don't take another one
-if [ -z `ls "$destination_dir" | grep $today_date` ]
-    then
+cd "$destination_dir"
+if [ -z "$(echo * | grep "$today_date")" ]; then
     mkdir -p "$backup_path"
-    rsync -aXh --stats --delete $no_perms "$source_dir/" --link-dest "$latest_link" --exclude-from="$HOME/.config/home-snapshot-excl.conf" "$backup_path" 2> $log_location
+    rsync -aXh --stats --delete "$source_dir/" $no_perms --link-dest "$latest_link" \
+    --exclude-from="$HOME/.config/home-snapshot-excl.conf" $exclude_relative "$backup_path" 2> $log_location
+    # Act depending on rsync exit
+    rsync_exit=$?
 else
     echo "Snapshot was already taken: $today_date"
 fi
 
-# Act dependig on rsync exit
-rsync_exit=$?
 if [ $rsync_exit -ne "0" ] && [ $rsync_exit -ne "24" ]
 then
     notify-send -a "rsync" -u critical 'Home snapshot possibly failed' "Check the log at $log_location, the service will retry in 10 minutes."
@@ -58,10 +70,11 @@ ln -rs "$backup_path" "$latest_link"
 
 #Cleanup old backups (keep only $max_snapshots)
 backups=$(ls -rt "$destination_dir")
-oldest_backup=$(echo "$backups" | awk 'NR==1{print $1}')
+oldest_backup=$(echo "$backups" | awk 'NR==1 {print $1}')
 let backup_count=$(echo "$backups" | wc -w)-1
-if [ $backup_count -gt $max_snapshots ] && [ -n $oldest_backup ]
+if [ "$backup_count" -gt "$max_snapshots" ] && [ -n "$oldest_backup" ]
 then
     echo "Deleting snapshot directory $destination_dir/$oldest_backup"
-    rm -rf "$destination_dir/$oldest_backup"
+    # block the command if destination_dir is not set somehow
+    rm -rf "${destination_dir:?}/$oldest_backup"
 fi
